requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
agno==1.2.4
agno-aws==0.0.1
agno-docker==0.0.1
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.21
botocore==1.37.21
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.16
duckdb==1.2.1
duckduckgo-search==7.5.4
email-validator==2.2.0
exa-py==1.9.1
fastapi==0.115.12
fastapi-cli==0.0.7
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
mako==1.3.9
mapepire-python==0.2.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.5.0
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.32.0
nest-asyncio==1.6.0
numpy==2.2.4
ollama==0.4.7
openai==1.68.2
packaging==24.2
pandas==2.2.3
pep249abc==0.0.1b3
pgvector==0.4.0
pillow==11.1.0
pluggy==1.5.0
primp==0.14.0
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
pyarrow==19.0.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pydeck==0.9.1
pyee==13.0.0
pygments==2.19.1
pypdf==5.4.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.24.0
s3transfer==0.11.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
streamlit==1.44.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.13.0
typing-inspect==0.9.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
